visual

+++ public/expenses/analytics/analytics-expenses-ctrl.js
@@ -16,12 +16,12 @@ angular.module("app").controller("analyticsExpenses", ["$scope", "$http", functi
                     if(i==response.data.length){
                     var data={
                         name: response.data.map(function(d) { return d["country"] })[i]+" "+ response.data.map(function(d) { return d["year"] })[i],
-                        data: parseFloat(response.data.map(function(d) { return d["countryExpense"] }))[i],sliced: true,
+                        data: parseInt(response.data.map(function(d) { return d["countryExpense"] }))[i],sliced: true,
             selected: true
                     };}else{
                          data={
                         name: response.data.map(function(d) { return d["country"] })[i]+" "+ response.data.map(function(d) { return d["year"] })[i],
-                        data: parseFloat(response.data.map(function(d) { return d["countryExpense"] }))[i]};
+                        data: parseInt(response.data.map(function(d) { return d["countryExpense"] }))[i]};
                     };
                     expensesData.push(data);
                     }
